# docs/swagger.yaml
definitions:
  auth.ReqDHParamsRequest:
    properties:
      a:
        type: string
      messageId:
        type: integer
      nonce:
        type: string
      serverNonce:
        type: string
    type: object
  auth.ReqDHParamsResponse:
    properties:
      b:
        type: string
      messageId:
        type: integer
      nonce:
        type: string
      serverNonce:
        type: string
    type: object
  auth.ReqPQRequest:
    properties:
      messageId:
        type: integer
      nonce:
        type: string
    type: object
  auth.ReqPQResponse:
    properties:
      g:
        type: integer
      messageId:
        type: integer
      nonce:
        type: string
      p:
        type: string
      serverNonce:
        type: string
    type: object
  biz.GetUsersRequest:
    properties:
      authKey:
        type: string
      messageId:
        type: integer
      userId:
        type: integer
    type: object
  biz.GetUsersResponse:
    properties:
      messageId:
        type: integer
      users:
        items:
          $ref: '#/definitions/biz.User'
        type: array
    type: object
  biz.GetUsersWithSqlInjectRequest:
    properties:
      authKey:
        type: string
      messageId:
        type: integer
      userId:
        type: string
    type: object
  biz.Sex:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Sex_Man
    - Sex_Female
  biz.User:
    properties:
      age:
        type: integer
      createdAt:
        type: string
      family:
        type: string
      id:
        type: integer
      name:
        type: string
      sex:
        $ref: '#/definitions/biz.Sex'
    type: object
info:
  version: 1.0.0
  title: 'Gateway Server'
  contact: {}
paths:
  /auth/req_dh_params:
    post:
      consumes:
      - application/json
      description: Request Diffie-Hellman parameters from server
      parameters:
      - description: Request Params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/auth.ReqDHParamsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ReqDHParamsResponse'
        "400":
          description: Bad Request
      summary: Request DH Params
      tags:
      - Auth
  /auth/req_pq:
    post:
      consumes:
      - application/json
      description: Request PQ from server
      parameters:
      - description: Request Params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/auth.ReqPQRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ReqPQResponse'
        "400":
          description: Bad Request
      summary: Request PQ
      tags:
      - Auth
  /biz/get_users:
    post:
      consumes:
      - application/json
      description: Retrieves a list of users from the database
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/biz.GetUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/biz.GetUsersResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get users
  /biz/get_users_with_sql_inject:
    post:
      consumes:
      - application/json
      description: Retrieves a list of users from the database using a SQL injection
        vulnerable query
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/biz.GetUsersWithSqlInjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/biz.GetUsersResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get users with SQL injection vulnerability
swagger: "2.0"